# Optimize varnish ban regexp