@bastnic 2021-04-21
@dunglas FYI xkey is already abandoned by varnish which prefers ykey https://docs.varnish-software.com/varnish-cache-plus/vmods/ykey/.

Migrating to x/ykey is not "free" as you need the paid version of varnish (maybe I'm wrong on this one).

This PR an optimization for what we already provide in apip.