@dunglas 2021-04-26
Yes xkey isn't maintained by Varnish.com (the company) anymore but is still maintained by the community as part of the Varnish modules project (I recently contributed to it and the patch has been merged). 
What I'm suggesting is to support both the current approach and xkey (that have better performance).